boiler_plate/requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-rest-framework==0.1.0
djangorestframework==3.13.1
passlib==1.7.4
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
